Allow [DataConverter] with SourceType of T on Property of type T
Concretely: a [DataConverter] whose SourceType is int should be able to be placed on a property whose CLR type is int?.
This is because nulls are filtered out before being sent to a DataConverter (that is, null is never converted and is
idempotently null). The bug was in the type cohesion check, which previously did not account for this scenario. The
existing unit test was insufficient because it used string?, which is indistinguishable from string in the type system.