Add object.ForDisplay Extension Method
This commit adds an extension method that can be used to convert a possibly null object into a string suitable for
display to users in error messages. The primary use case is to include the arguments passed to an attribute such as
[Check] in the error message when an error occurs during construction. For this purpose, we want to distinguish strings
with double quotes and characters with single quotes; plus, we want to recursively handle arrays. This extension will
also be useful when translating other annotations that take arguments, such as the Ranged Comparison annotations and the
Inclusion Annotations.