Complete Refactoring of Translation Layer
This commit fully refactors the Translation Layer of Kvasir, including a rewrite of the majority of the system's
unit tests. There were several reasons for this, mostly stemming from forward compatibility. Although the original
implementation was functional, it was organized in such a way that would have either precluded future Translation
(e.g. of Aggregates) or would have required a massive rewrite anyway. The new design, entirely in the impleentation
details, is deliberately cognizant of these future needs.

The new implementation relies on a "base translation" that is repeatedly updated as new information is extracted
from annotations and contexts. These base translations make their way through the system as a mapping keyed on the
source access path, making it trivial to resolve 'Path' properties on annotations. This pattern makes it so that
the majority of the Translation Layer does not have to differentiate between e.g. Scalars and Aggregates - the former
is simply a mapping of size 1 with the empty string as its sole key. The expectation is that this design will make
it fairly straightforwad to implement the remaining four categories of Translation (namely: Aggregates, Enumerations,
Relations, and References).

In addition to refactoring the Translation code itself, a number of specification changes were made. For example,
redundant annotations are no longer errors; instead, they just have no effect. There is also a new system by which
constraints are flattened, ensuring that the final set of constraints applied to a Field is minimal. Inherited
properties can also now be marked as [IncludeInModel], though they are excluded by default. Error messages are more
standardized, and a larger number of scenarios are actively tested (e.g. interplay between annotations, null values
being provided, etc.).

As part of the refactor, a few changes were made to some pieces outside of the Translation Layer, mostly in the
annotations space. These changes were largely aesthetic: nothing about the user-facing APIs changed, but certain
inheritance hierarchies were modified to reflect the relationships that are actually relevant to Translation. The
object.ForDispaly extension method was also updated to handle Enumerations and to better handle edge cases with
integer-valued floating-point numbers.