Translation of Enumeration-Type Fields
This commit adds support for translation of enumeration-type Fields. For the most part, this translation is the same as
translating a scalar: names, nullability, and default values are processed the same way. The biggest difference is that
enumeration-type Fields have an implicitly restricted image that is not recognized as a constraint, but instead as a
special property of the EnumField in the Schema Layer. The discreteness constraints can still be applied, but they
operate on the restricted image. Other constraints, aside from IsNot, are inapplicable.

There are also two new annotations that are valid only on enumeration-type Fields: [IsString] and [Numeric]. These
annotations direct the framework to use either the string or underlying numeric representation, respectively, as the
native storage in the back-end database, even if that back-end database has support for enumerations directly. From a
technical perspective, these are equivalent to a DataConverter annotation with a pre-defined converter type.