Utility for Obtaining Array of Valid Enumerators
This commit adds a utility function that, given an Enumeration type, produces an unordered array of the type's valid
enumerators. This is a companion function to the long-existing Enumerator.IsValid() extension method, and has the same
general-purpose use cases. Specifically, though, this function will be used by the Translation Layer when translating
Enumeration-type Fields, as the Schema Layer has to be seeded with the allowed set. Note importantly that this new
method is computationally expensive for flag enumerations, and therefore the IsValid() extension does not simply call
this method and then check for containment.