Translation of Aggregate Fields
This commit implements the full solution for translating Aggregate-type Fields. There are a number of differences from
how Scalars and Enumerations are translated, and a number of new algorithms and patterns had to be implemented to
elegantly support both models. Aggregates required no new annotations to translate, and no existing test cases had to be
modified.