Make Field Name a List of Parts During Translation
This commit changes the way that a Field's name is tracked during Translation. Previously, the FieldDescriptor stored a
single string; this was sufficient for Scalars and Enumerations, because there is only ever one part. However, with
Aggregates, there may be multiple parts: specifically, one part per nested Aggregate (e.g. Foo.Bar.Baz, which has three
parts). It is important to be able to identify which substring belongs to which part, as Field renaming with Aggregates
and nested Fields is complicated and relies on overwriting sections of a name. A single string is insufficient because a
client can manually rename a Field to be a string that contains a dot (the natural separator), which would introduce
ambiguity if not separated out independently.