Differentiate Between Absolute and Relative Columns in Translation
This commit changes the way that column indices are handled in Translation; specifically, the single optional Column
Field has been changed to a required RelativeColumn Field coupled with an optional AbsoluteColumn Field. The former
represents the relative offset of the Field within its group; this starts at 0 (with Scalars and Enumerations) and is
then updated after the column indices are assigned for a given type. This pattern allows us to differentiate between the
order required by an Aggregate and the order requested by the user: the previous pattern did not allow for this
disambiguation, and therefore left no way to forward the former information downstream.