Defer Analysis of Nullability to Annotation Processing
This commit defers the entirety of the analysis of nullability for a Field to the dedicated annotation processing
function. Previously, a default assessment was performed when the base translation was made. This works for Scalars and
for Enumerations, but with Aggregates it makes it impossible to handle compound nullability. Specifically, consider the
following hypothetical example:

   1. A nested Scalar/Enumeration Field is natively nullable (e.g. string?)
   2. The Aggregate property is also natively nullable
   3. The Aggregate property is annotated as [NonNullable]

With initial nullability being processed up front, there is no way to identify that the [NonNullable] annotation only
overwrites the nullability of the Aggregate property itself, resulting in the nested Field still being nullable in the
encapsulating Entity. We now process both the natively nullability and any nullability annotations simultaneously, and
make changes only when the combination calls for it.