Add AccessPath to FieldDescriptor
This commit simply adds the dot-separated access path for a Field to its FieldDescriptor. This is the same value that is
used as the key for the state that traverses through the Translation layer, but this will make certain algorithms
cleaner to write in the future. (Specifically, it avoid having to use tuple notation for (string Path, FieldDescriptor
Descriptor) or anything similar, including KeyValuePair<>.)