Ensure that Fields are Sorted by Column Index when Creating an Entity
This commit makes sure that the Fields for a type are sorted by column index when creating an Entity and populating its
Fields. We were getting away with this before because of the way that our various LINQ queries to convert from the
Translation state dictionary were operating, hence no new unit tests; however, I don't want to rely on this behavior,
hence the change.