Implement (But Do Not Yet Use) New Column-Solving Algorithm
This commit implements a new column solving algorithm that leverages the StickyList from Cybele but is also capable of
handling sequential groups. Cybele's StickyList only supports one element at a time, which is insufficient for the needs
of Aggregates. What we actually have is a NP-hard problem of identifying if we can create N subsets of a set of numbers
such that each subset's sum matches exactly one of a set of sums. (Basically, a more complex version of the standard
subset sum problem.) We solve this by brute force, given that I don't expect the number of Field groups we're dealing
with to ever be all that large. I'm using a NuGet package for permutation generation, simply because I don't feel like
implementing and testing my own algorithm.

Note that Translation does not yet use this new algorithm; it is pre-emptive in preparation for the Aggregation
translation, and has been separated into its own commit for ease of review and clarity of history.