Translation of Reference Fields
This commit implements the full solution for translating Reference-type Fields. Translating a Reference Field is
basically the same as translating an Aggregate Field, except that the nested contribution is only the Fields of the
Reference's Primary Key. It also requires the creation of ForeignKeys as part of the Table schema. References required
no new annotations to translate, and no existing test cases had to be modified.