Deduce Primary Key as All Fields if All Non-Nullable and No Other Deduction
Each row in a database Table must be unique: this follows directly from the fact that each row must have a unique
Primary Key that is a (possibly proper) subset of the row's Fields. As such, there is an implicit Candidate Key
consisting of all the Fields in a Table that should be considered when deducing a Table's Primary Key. This commit adds
that deduction opportunity. Specifically, the Primary Key of a Table will be deduced as the complete set of Field
comprising that Table if all the Fields are non-nullable and the deduction algorithm does not find another viable
Primary Key. Note that this Candidate Key will never actually appear in the final DDL, as it is (by defintion) a
(possibly proper) superset of the Primary Key and therefore filtered out.

While this is an appropriate change for all existing Translation vectors, it is particularly important for translating
Relation. Unlike regular Entities, the synthetic Entities that back a Relation Table have an implicit Primary Key of
"all the Fields." While there will be a mechanism to override this (namely, the [PrimaryKey] attribute with a <Path>) we
need a way to allow users to add no annotations. This is that way.