Upgrade to .NET7, C#11
This commit upgrades all projects to .NET7 and C#11. The only feature of C#11 currently being employed is static virtual
interface methods, which are in use for the ConnectionType property of the IRelation interface. (Note that there is a
minor unfortunate implementation caveat that is documented in-source related to the use of this feature.)