Introduce PropertyCatalog
This commit introduces the PropertyCatalog, which takes the place of the FieldListing as the active state-tracking
structure for Translation. A PropertyCatalog contains a FieldListing as well as a mapping of access paths to Relation
representations; that representation is currently typed as an object, as it is unused. There is no change to the actual
Translation logic, but this is necessary to enabel the tracking of Relation properties across function calls and types.