[RelationTable] Annotation for Renaming Relation Tables
This commit adds a brand new annotation, [RelationTable], that operates like the [Table] annotation for
Entities but can be applied to Relation-type properties. It's behavior is basically identical, changing
the default name of the Relation Table that backs the property. Only one is allowed per property, and
it cannot be nested via a path attribute.