ConnectionType on Read-Only Relations
This commit adds the ConnectionType static property to the read-only Relation interfaces. Users may choose to
type their Entities' properties using the read-only interfaces, in which case the reflection APIs that will
ultimately extract the ConnectionType will need a way to do so. Since the read-only interfaces have access to
the same generic information as the write-capable flavors, this is not an issue.