Dedicated Synthetic Name for UniqueAttribute
This commit adds a new reserved name for UniqueAttributes to be used for the natural uniqueness of the
(Entity, Key) pair for map-like Relations. This reserved name begins with the reserved anonymous prefix
but is suffixed by something that will never be obtained for a "regular" anonymous constraint. It is
not possible for a user to legally provide the reserved name to their own applied attribute; it is for
library use only.