Nestable Annotations
This commit slightly extends the Annotation type hierarchy by introducing a new interface that represents an attribute
that can be "nested" - that is, one that has a Path attribute. This new interface exposes that Path attribute and a
singular API that clones the attribute while providing a different Path. This will be necessary for translating Relations,
where the Path provided by the user to an annotation needs to be cleaved when considered for the Relation Table. By
integrating this into the attributes themselves, we keep the system extensible and don't have to do any pattern matching
or anything like that if we add new attributes.