RelationOrderedList
This commit adds the RelationOrderedList, a Relation that implements the same API as a List<T> but keeps track of the
position of its elements. This means that removing or inserting elements causes shifts, which are realized as
modifications to the items rather than a bunch of deletions; this is the only Relation that makes use of this status.

This commit only implements the collection itself. Support for translating RelationOrderedLists in the schema layer will
come in a subsequent commit.