Data Translation of RelationOrderedLists
This commit extends some of the existing tests, and adds two new tests, for RelationOrderedLists. These are translated
basically the same as RelationMaps, with the Fields being 'Index' and 'Item'. The only new tests are for things related
to keys (Primary Key identification and Candidate Key definitions), as the existing tests for other relations cover all
other cases. I've chosen not to retrofit any existing tests for things like constraints to be in terms of
RelationOrderedLists.