Fix Mis-Naming of Explicit Interface Implementations
This commit fixes a bug in the way that Fields backed by explicit interface implementations were translated. Previously,
the name of the Field defaulted to the fully-qualified name of the source interface plus a dot plus the name of the
property, as this is the value of the property's Name according to the reflected PropertyInfo. This is obviously
undesirable, so we now grab only the last portion of the dot-separated Name exposed by the PropertyInfo. Note that this
can lead to implicitly duplicated Field names, since two different interfaces can define properties with the same name
that are then both implementated; resolving this with a [Name] annotation is the responsibility of the user.

(Note that this bug affected all five varieties of property, but the unit test was written for scalars only for
simplicity.)