Allow Extraction and Reconstitution to Work with Enumerations
This commit fixes bugs in both the Extraction and Reconstitution layer related to the when data is viewed as a
possibly-null object and when it is viewed as a DBValue. The key is that the latter is incapable of representing a C#
enumeration, since that is not something that literally exists in the back-end database. Previously, primitives were
extracted as objects and immediately wrapped in DBValue; they were then unwrapped to do data conversion before being
re-wrapped. This led to a crash if an enumeration were extracted. (The same happened when reconstituting, just in the
opposite direction). The solution was to make everything operate in terms of raw objects until after data conversion (or
until data reversion) occurs, when it is safe to use DBValues.