Properly Detect and Report Attempted Translation of Static Class
Static Classes in C# are implemented in the IL as abstract and sealed, a combination which is otherwise not supported.
We were previously detecting an attempted translation of a Static Class and erroring out with a message indicating it
was abstract; this commit makes it so that such a case is identified in the error message more precisely.