CurryingExtractor
This commit adds a new IMultiExtractor implementation, the CurryingExtractor. This implementation serves a similar niche
as the DecomposingExtractor, and is designed to work in tandem with it. The DecomposingExtractor works great for running
extraction over an Aggregate, but it expects each of the sub-extractors to accept the same type. When an Aggregate is
nested in another Aggregate, that becomes a problem, because what we really want is to first pull the nested Aggregate
out and then run that Aggregate's extraction logic. That is what the CurryingExtractor does.