Fix Source and Result Types of ISingleExtractor Implementations when Nullables are Involved
This commit fixes a bug whereby the ReadPropertyExtractor and the ConvertingExtractor were not properly unwrapping
Nullable<T> to the underlying type when Nullable<T> was either the Source or Result types. Because Extraction always
operates in terms of object?, the boxing/unboxing renders the Nullable<T> wrapper useless. Furthermore, by unwrapping
the type for the SourceType and ResultType properties, it makes other debug checks much easier.