Fix Mistranslation when DataConverter Result Type is Enumeration
This commit fixes a bug in the basic Translation layer that mishandled [DataConverter] annotations where the resutl type
was an enumeration. Previously, the restricted image portion of the FieldDescriptor was not updated, leading to zero
valid enumerators being available when the EnumField instance was created. This has been fixed, with the system now
correctly identifying scalar-to-enumeration and enumeration-to-enumeration conversions.