Rewrite of Schema Translation (Again)
This MASSIVE commit is a COMPLETE refactor of the Translation Layer. While working on the logic for translating
extraction plans, I realized that the existing implementation -- heavily dependent on simple aggregate structs,
mostly functional flow, and string-based maps -- was too fragile to really support extraction. To that end, I
took the opportunity to rethink the way that I had designed the Translator class (again), and I believe that the
new design is much stronger, less brittle, and more future-proof. A lot (nearly everything) has changed, but the
main pieces are outlined below.

Descriptors and Groups
======================
The new Translator design is based on two fundamental building blocks: FieldDescripors and FieldGroups. A
FieldDescriptor describes a single Field coming from a single scalar property, while a FieldGroup describes one
or more properties that are grouped (e.g. from an Aggregate or a Relation). These types form inheritance
hierarchies, which allow for specialization of annotation handling and easier error checking: for example, the
base FieldDescriptor does not allow for comparison CHECK constraints, but the derived OrderableFieldDescriptor
does. When a FieldDescriptor or a FieldGroup is constructed, it performs reflection to handle all the annotations
on the source proeprty; for FieldGroups, this also means decomposing paths. The key to FieldGroups is that they
can encapsulate other FieldGroups: each level is implemented as a simple mapping of the _next_ path segment to the
FieldGroup at that segment. This makes dealing with nested properties so much cleaner.

Contexts
========
As translation progresses, we now use a Context object to track where we are and to then contextualize error
messages. A Context keeps track of the full chain that got us to the current property or type and can regurgitate
that when requested. This means that Context objects are passed around heavily, but it is much cleaner than using
string paths or other tracking. Furthermore, because the Context knows where we've already been, cycle detection
is trivial.

Derived Error Classes
=====================
The new Translation subsystem has tens of derived Exception classes for different categories of errors, rather than
a single exception-creating helper unit. While this may be overkill, it makes it far easier to quickly identify the
problem from a line of code, and it mmakes it possible for unit tests to expect a specific kind of exception. The
error message format is also standardized, following the same Location/Path/Problem structure but in a formalized
capacity. At the same time, there are some new unit test extensino methods to look for those lines specifically.

For the most part, everything else is generally the same aside from some implementation details. We still use
Synthetic Types for dealing with Relations, but there is now a dedicated function for creating them. We still defer
translation of Relations, but we handle the initial annotations with a new RelationTracker class for in situ
processing. Nothing about the assignment of columns or the deduction of Primary Keys has changed, though the
implementation and documentation has been cleaned up a little.