Add New Schema Translation Tests
This commit adds a number of tests of the Translation Layer's schema functionality that were previously missed. Many of
these cases were identified during the planning for the complete refactor of that subsystem, though others were just
identified via inspection. Note that these are all tests that cover missing logic, not missing lines of code, so it
should have no impact on code coverage.

In particular, the new tests cover:

    * Column collisions identified in an Aggregate (rather than in an Entity)
    * Constraints that are modified by upstram annotations and applied via Paths
    * Constraints on Primary Key fields of References that should not propagate to the foreign keys
    * De-duplication of identical [Name] annotations
    * De-duplication of Primary and Candidate Key annotations from upstream applied via Paths
    * Default values that are originally valid and are later invalidated by upstream constraints
    * A few other one-off scenarios