Fix Incorrect Displays for Infinity, Scientific Notation Floats
This commit fixes a bug that caused positive/negative infinity and all floats represented in scientific notation to be
displayed with a trailing .0. This was happening because the values would get caught up in the non-integral branch,
which seeks to append the trailing .0 for floating-point values that are equivalent to integers. The fix is to have a
separate carve out for these two sub-cases.