Remove Translation v1, Replace with Translation v2
This commit DELETES all of the Translation v1 code and then renames the Translation v2 subdirectory and
namespace to drop the v2 suffix. This is done as a standalone commit because the diff may be a bit ugly,
with some files appearing to be brand new and other files appearing to have been modified.