Tests for Uncovered Code and Unconsidered Scenarios
This commit adds a handful of tests and makes some very small code changes to improve the overall code
coverage of the repository and address some scenarios that were previously untested. The primary goal
of this commit is to ensure that the Translation Layer's schema translation logic is thoroughly unit
tested and as foolproof as possible.