Populate Restricted Domain when Boolean Property is Data Converted
This commit adds logic to populate the restricted domain of a Field when the source type is a Boolean but a data
conversion is applied to make the Field's type something different. We can do this because we know that the only valid
CLR values for the property are 'true' and 'false', so they get fed through the data converter. We already did this when
a Boolean field was converted to a Boolean, and when dealing with conversions from Enumerations.