Disallow By-Ref Properties
This commit disallows properties whose return type is by-ref from being included in the Kvasir data model. This required
a new TypeCategory and a new branch in the type-display logic, but otherwise was a vanilla change. These properties are
disallowed because I don't know what their implications are for data changing, so I'd rather disallow them for now; I
can always loosen that restriction in the future. (But honestly, I probably won't.)