Translation of Extraction Plans
This commit implements the Translation layer for Extraction, which is the process by which data is pulled out of CLR
objects and into 'rows' for the database. Annotations do not apply to the Extraction layer, and there are no errors
possible. All classes of property are supported: scalars, enumerations, aggregates, references, and relations (including
arbitrary nesting).