Populate EntityLookup Function on Constructon of Translator
This commit makes it so that an Entity Lookup function is required to create a Translator. In practice, this will be
sourced from an EntityDepot in the same scope as the Translator, and the lookup function will then be used to drive
reconstitution of Reference-type fields. None of the existing tests actually need this function, but I don't want to
have a default constructor just for convenience in unit tests, so all the unit tests have been updated to pass a no-op
function instead.