Allow KeyLookupCreator to Handle Null Keys
This commit updates the KeyLookupCreator to return a null result if the key being looked up is all nulls. This is to
support nullable references, where the values passed to the ICreator interface are indeed all null. The original design
had the caller checking for nulls before invoking the API, but that breaks the generalization.