Implement DataReconstitutionPlan in terms of ReconstitutingCreator
This commit changes the implementation of the DataReconstitutionPlan to be purely in terms of the ReconstitutingCreator
implementation of the ICreator interface. Previously, the DataReconstitutionPlan was highly flexible, allowing for any
ICreator and a list of IMutators. This flexibility is unnecessary, though, because we will always be using a
ReconstitutingCreator; in fact, those are the same arguments as are used for the ReconstitutingCreator constructor.

Since this type is not yet in use in the framework, only unit tests were affected.