Ensure that DefaultStructCreator can Produce Null from Row of All Nulls
This commit extends the behavior of the DefaultStructCreator to match that of the ConstructingCreator when dealing with
a row of all null values. Specifically, the DefaultStructCreator now takes an argument controlling whether or not a type
can be created from a row of all null values. Even though the arguments won't be passed to the default constructor
(obviously), we still have to handle this case to avoid creating an object in circumstances when we want to have a null
value.