Transaction Commands Interfaces
This commit adds two interfaces: ICommands and ICommandsFactory. The former is responsible for producing DBCommands for
CRUD operations against a particular Table, and the latter is responsible for creating the former. Conceptually, an
implementation of the ICommandsFactory interface will correspond to a particular back-end database system and/or SQL
flavor. These are the APIs that the Kvasir framework will interact with to load and store data, enabling support for
arbitrary RDBMSes.