Transactor: CreateTables
This commit defines the initial shape of the Transactor class and implements the table creation API. The Transactor is
the component that is responsible for bridging the gap between the C#/CLR world and the database world. It uses various
interfaces to generate the appropriate SQL commands for a given Table (that is, the Transactor knows nothing about the
actual logic required to do something) and ensures that execution is properly ordered, formally transacted, etc.

For creating tables, the logic is fairly straightforward. The most complicated piece is the topological sort required to
ensure that dependent Tables are created before depending Tables. Principal Tables are always created first, followed by
all Relation Tables. Tables are created as a single atomic transaction, and we are able to handle both failed commits
(via rollback) and failed rollbacks.

For testing, a custom fixture has been created that sets up all the necessary mocks, since this is a very mock-heavy
scenario. The use of interfaces throughout the Transactor allows for seamless use of arbitrary back-end RDBMS systems
(including custom implementations like JSON dumping) without altering the core business logic. A nice side effect of
this design is easy mockability via dependency injection.