Transactor: SelectAll
This commit defines the logic for performing the 'SELECT ALL' query against the back-end database. This query grabs all
of the data from all of the tables (Principal and Relation) and then runs the Reconstitution and Repopulation Plans to
turn that data back into CLR objects. The selections are done according to the topological ordering, with all Relations
being selected-from after all Principal Tables. The KeyLookup logic initially implemented for the Reconstitution Layer
is reused to match rows of relation data to the appropriate owning Entity.