Add KeyExtractor to Translation of Principal Table
This commit adds the KeyExtractor to the EntityTranslation for the Principal Table. This is something we already had
available in the code, sice we use it to create the KeyMatchers for key-lookup reconstitution, but we weren't saving it
anywhere. The KeyExtractor will be useful when implementing the 'SELECT ALL' query in the Transaction layer, since
matching relation rows to their owning Entity is functionally identical to reconstituting a Reference-type field.