Transactor: Insert
This commit defines the logic for performing an 'INSERT' query against the back-end database for one or more entities.
This query handles the data for the Principal Table and any Relation Tables, ensuring that everything is ordered
correctly so as not to induce Foreign Key constraint violations. Failed inserts (e.g. because of duplicate Primary Keys)
are handled with transactional rollback.