Add Canonicalization Function to RelationExtractionPlan
This commit adds a function to the RelationExtractionPlan that takes a CLR source object, extracts the target Relation,
and then invokes the Relation's canonicalization method. This is needed in the Transaction Layer: after performing a
database write (insert or update), we need to reset the Relation's tracking information. This cannot be done as just a
postscript to the extraction logic because the database write might fail, in which case the tracking should not actually
be reset.