Transactor: Delete
This commit defines the logic for performing a 'DELETE' query against the back-end database for one or more entities.
This query handles the data for the Principal Table and any Relation Tables, ensuring that everything is ordered
correctly so as not to induce Foreign Key constraint violations. Failed deletions (e.g. because of database errors) are
handled with transactional rollback.