Transactor: Update
This commit defines the logic for performing an 'UPDATE' query against the back-end database for one or more entities.
This query handles the data for the Principal Table and any Relation Tables, which includes executing INSERTS and
DELETES for Relation Table rows as appropriate. There is a guaranteed ordering, but it's not strictly necessary, as it
is assumed that the Primary Key for a row never changes. Failed updates (e.g. because of illegal Foreign Keys) are
handled with transactional rollback.