Add Error Check for Writeable Relation Properties
This commit adds a new Translation error that is raised when a Relation-type property that is included in the data model
is freely writeable. Notably, this excludes init-only properties. The intent here is to prevent properties that allow
users to freely change the actual relation object rather than just manipulating the collection's contents: the former
causes issues with state tracking, while the latter is obviously fine.