Fix Bug that Missed Detecting Nested Relations due to Memoization
This commit fixes a bug in the Translation layer whereby a certain layout led to a Relation-nested Relation not being
flagged as illegal. Specifically, the bug was triggered by the following set-up:

   1. A Relation is nested in an Aggregate 'A'
   2. A Field of type 'A' is translated successfully
   3. A Relation-type Field that has 'A' as an element is translated

Because the result of (2) is memoized, when 'A' is re-encountered during (3) we previously skipped all additional
translation and simply returned back the all the non-Relation Fields of 'A'. This memoization causes the discrepancy: if
(2) had never happened, then (3) would have correctly identified the illegal schema.

The solution is to detect when something like this is going to happen (thanks to RelationTrackers) and simply pretend
that the type's translation is not memoized. In letting the normal course of translation continue, the regular detection
path will be executed and the correct (contextually rich) exception will be raised. We know that no other errors can be
raised, since the type was translated successfully once already.

A new unit test, which failed to trigger an exception before the change, has been added.