Fix Bug causing Missing Information in Error Messages for Nested Relations
This commit fixes a bug that led to missing path information in error messages contextualizing Relations nested within
Aggregates. The previous code implicitly assumed that the Relation was always on the top-level Entity, and could
therefore get the first portion of the SyntheticType's typename from the reflected type of the RelationTracker's
property. But this was wrong if the Relation was actually nested within an Aggregate; in that case, the reflected type
was the Aggregate, and the Entity type (and any intermediate pathing) would be lost.

To solve this, I had to rework the RelationTracker a bit. A small amount was quality-of-life (specifically, the changs
to the AnnotatedName) but the meat (specifically, extending via property and tracking both contextualized and
decontextualized paths) were strictly necessary. I also moved some logic from other parts of the code base into the
RelationTracker to avoid breaking encapsulation.